move pollagent to client side to allow toasts in calbacks
